Fixed up the namespace
From what I can see, at least when compared to other laminas and Mezzio
packages, the previous namespace was unnecessarily convoluted. This
change simplifies it, and reorganises the files and directories to match
it.

The change also cleans up the files, removing unnecessary methods, and
class member variables.
