+++ test/ConfigProviderTest.php
@@ -5,6 +5,8 @@ declare(strict_types=1);
 namespace MezzioTest\StaticPages;
 
 use Mezzio\StaticPages\ConfigProvider;
+use Mezzio\StaticPages\Handler\StaticPagesHandler;
+use Mezzio\StaticPages\Handler\StaticPagesHandlerFactory;
 use PHPUnit\Framework\TestCase;
 
 /**
@@ -14,7 +16,21 @@ class ConfigProviderTest extends TestCase
 {
     public function testHasDependencies(): void
     {
-        $configProvider = new ConfigProvider();
-        $this->assertArrayHasKey('dependencies', $configProvider());
+        $this->assertArrayHasKey(
+            'dependencies',
+            (new ConfigProvider())()
+        );
+    }
+
+    public function testCanGetDependencies(): void
+    {
+        $this->assertEquals(
+            [
+                'factories' => [
+                    StaticPagesHandler::class => StaticPagesHandlerFactory::class,
+                ],
+            ],
+            (new ConfigProvider())->getDependencies()
+        );
     }
 }
