+++ test/Handler/StaticPagesFactoryTest.php
@@ -2,16 +2,15 @@
 
 declare(strict_types=1);
 
-namespace Settermjd\StaticPages\Test\Handler;
+namespace MezzioTest\StaticPages\Handler;
 
 use Mezzio\Exception\MissingDependencyException;
-use Mezzio\Router\RouterInterface;
+use Mezzio\StaticPages\Handler\StaticPagesHandler;
+use Mezzio\StaticPages\Handler\StaticPagesHandlerFactory;
 use Mezzio\Template\TemplateRendererInterface;
 use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
 use Psr\Container\ContainerInterface;
-use Settermjd\StaticPages\Handler\StaticPagesHandler;
-use Settermjd\StaticPages\Handler\StaticPagesHandlerFactory;
 
 class StaticPagesFactoryTest extends TestCase
 {
@@ -23,17 +22,19 @@ class StaticPagesFactoryTest extends TestCase
         $this->container = $this->createMock(ContainerInterface::class);
     }
 
-    public function testFactoryWithRouteAndTemplate(): void
+    public function testFactoryWhenContainerHasATemplateRendererInterface(): void
     {
-        $router            = $this->createMock(RouterInterface::class);
         $templateInterface = $this->createMock(TemplateRendererInterface::class);
 
         $this->container
+            ->expects($this->once())
             ->method('has')
-            ->willReturnOnConsecutiveCalls(true, true);
+            ->with(TemplateRendererInterface::class)
+            ->willReturn(true);
         $this->container
             ->method('get')
-            ->willReturnOnConsecutiveCalls($router, $templateInterface);
+            ->with(TemplateRendererInterface::class)
+            ->willReturn($templateInterface);
 
         $factory = new StaticPagesHandlerFactory();
 
@@ -42,34 +43,16 @@ class StaticPagesFactoryTest extends TestCase
         $this->assertInstanceOf(StaticPagesHandler::class, $page);
     }
 
-    public function testFactoryWithoutRoute(): void
-    {
-        $this->expectException(MissingDependencyException::class);
-        $this->expectExceptionMessage("RouterInterface object not found in the container");
-        $this->container
-            ->method('has')
-            ->with(RouterInterface::class)
-            ->willReturn(false);
-
-        $factory = new StaticPagesHandlerFactory();
-
-        $page = $factory($this->container);
-
-        $this->assertInstanceOf(StaticPagesHandler::class, $page);
-    }
-
-    public function testFactoryWithRouteButWithoutTemplate(): void
+    public function testFactoryWhenContainerDoesNotHaveATemplateRendererInterface(): void
     {
         $this->expectException(MissingDependencyException::class);
         $this->expectExceptionMessage("TemplateRendererInterface object not found in the container");
 
-        $router = $this->createMock(RouterInterface::class);
         $this->container
-            ->method('get')
-            ->willReturn($router);
-        $this->container
+            ->expects($this->once())
             ->method('has')
-            ->willReturnOnConsecutiveCalls(true, false);
+            ->with(TemplateRendererInterface::class)
+            ->willReturn(false);
 
         $factory = new StaticPagesHandlerFactory();
